New functionality: account can send email from a domain with any address
Originally during the sending of email there is a check to not send out
email from an unaothorized address. There you can set only exact email
addresses.
With this modification you are able to send out any address within a
specified domain. This is the reverse methodology to the catchall
function, you can use this to catch everything and send everything in a
domain, usable for a project management/ERP software.

To enable this you need to set "@domain.com" as sender address alias
just like you set up a cacthall functionality.

Signed-off-by: Csaba Tóth <csaba.toth@i3rendszerhaz.hu>