FIX: relaydomain MX list should matched with the actual domain's MX list
If you create a relaydomain, than the DKIM and SPF cgecks are useful,
because they check that in the domain's settings you really have good
config. But the MX check is not useful, because it tries to match with
your valid MX list (setted globally in the config panel), what is a good
behavior for a handled domain, but for a relay domain it is not.

This modification modifies the MX record checking methodology that it
checks if your relay hop is setted as a MX receiver for that domain, so
you can check that you forward your emails into the good target server
or not.

Signed-off-by: Csaba Tóth <csaba.toth@i3rendszerhaz.hu>