FIX: good floating IP MX record check for a relay domain
Google have a floating IP solution for load balancing, if you query the
aspmx.l.google.com address you get another IP in every second. Because
of this the MX check fails, because the original code checks only that
you have the same IP address in the list.
With this modification if the FQDN matches, the system validates the MX.

Signed-off-by: Csaba Tóth <csaba.toth@i3rendszerhaz.hu>