Bump djangorestframework to 3.11 and move to OpenAPI
See: https://www.django-rest-framework.org/community/3.10-announcement/

CoreAPI is being deprecated and since 3.10, settings must be modified to
keep using CoreAPI schemas. This is especially required for the
documentation generation.

To prevent settings modification while upgrading Modoboa and to follow
the standard for modeling Web APIs, CoreAPI is removed. Instead, OpenAPI
is used to generate the schema and Swagger UI is used to render the
documentation as described here:

  https://www.django-rest-framework.org/topics/documenting-your-api/