Bump Django to 2.1 and rename view permissions
See: https://docs.djangoproject.com/en/stable/releases/2.1/

The main change is the addition of a view permission. To prevent
redundancy, the custom and current `view_` permissions are renamed to
use the new ones - e.g. `view_aliases` is renamed to `view_alias`. Only
`view_domain` is used to check permissions and `view_domains` is not
deleted in case customizations have been made by user - in such case
a manual squash is needed.