Raise domain name length to RFC1035 maximum
RFC1035 Section 2.3.4. specifies a maximum domain name length of 255
octets in protocol. As each label is prefixed by its length and the
domain-root (.) has a length of 0 we have a maximum of 253 octets. For
modoboa to support all these domains in its DB scheme we need to raise
the maximum domain name length to 253 as well.