Handle NameTooLong by logging an error
get_dns_records() could be fed with a name that is too long for the DNS
protocol. In this case a NameTooLong error is thrown, which we can catch
and log an error for. This can happen for example if a DNS name that is
already at the limit is used in combination with "manage.py modo
check_mx", which appends an autodiscover to the domain, which then
exceeds the DNS name limit.