Fix for autoreply when timezone offset is - from UTC instead of +
Current autoreply code assumes the timezone offset from UTC always has a + character in front of it. This does not work for timezones in the western hemisphere as the offset has a - in front, and results in an internal error from web interface.

Instead split it at 6th character from right, so it works regardless of + or - offset.